backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
plotly
numpy
pydantic-settings
python-dotenv
python-multipart
google-generativeai
langsmith
matplotlib
